+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+
+### Added
+
+- Gulp-inject and Bower-installer to inject JavaScript dependencies direct to `functions.php`.
+
+### Changed
+
+- Bower folder name to `vendors/`.
+
+## [1.2.0] - 2015-11-02
+
+### Changed
+
+- All php files updated with latest version of `_s`.
 
 ## [1.1.0] - 2015-09-22
 
@@ -11,7 +28,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - `sass/` folder renamed to `scss/`.
-- Gulp replaced Grunt.
+- Switched to Gulp.
 - Removed mixins on main layout and sidebar, added classes to markup instead.
 
 ### Removed
@@ -29,6 +46,3 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Modifying basic HTML structure.
-
-[unreleased]: https://github.com/thinkxl/whiteboard/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/thinkxl/whiteboard/compare/v0.1.0...HEAD